Clean up project structure, added new services
