+++ .gitignore
@@ -2,5 +2,18 @@ env
 checkpoints
 .ideaenv
 __pycache__
-postgresql-42.7.1.jar
-data-postgres
+
+
+voting_service/__pycache__
+voting_service/db_operation/__pycache__
+
+spark_service/__pycache__/checkpoints
+spark_service/postgresql-42.7.1.jar
+
+realtime_dashboard_service/__pycache__
+
+postgres_service/.postgres_service
+
+ingestion_service/__pycache__
+ingestion_service/env
+ingestion_service/db_operation/__pycache__
