+++ ingestion_service/db_operation/create_tables.py
@@ -68,10 +68,4 @@ def insert_candidates(conn, cur, candidate):
         candidate['campaign_platform'], candidate['photo_url']))
     conn.commit()
 
-def insert_votes(conn, cur, vote):
-    cur.execute("""
-            INSERT INTO votes (voter_id, candidate_id, voting_time)
-            VALUES (%s, %s, %s)
-        """, (vote['voter_id'], vote['candidate_id'], vote['voting_time']))
 
-    conn.commit()
