+++ ingestion_service/main.py
@@ -1,55 +1,36 @@
-from data_src.create_tables import insert_voters, insert_candidates, create_tables
-from data_src.generate_data import create_candidate_data, generate_voter_data, trump_data, harris_data
 import psycopg2
-from confluent_kafka import SerializingProducer
 import simplejson as json
-from data_src.connect_postgres import connect, cursor, get_candidates
-from config import KAFKA_CONFIG, kafka_producer_config
+from confluent_kafka import SerializingProducer
+from kafka_conn import kafka_producer
+from db_operation.connect_postgres import connect, cursor, get_candidates
+from config import KAFKA_CONFIG, kafka_producer_config, NUMBER_OF_VOTERS
+from db_operation.create_tables import insert_voters, insert_candidates, create_tables
+from db_operation.generate_data import generate_voter_data, add_candidates, trump_data, harris_data
 
+"""
+create candidates and insert voters into postgres  
 
-
-
-# Kafka Topics
-def delivery_report(err, msg):
-    if err is not None:
-        print(f'Message delivery failed: {err}')
-    else:
-        print(f'Message delivered to {msg.topic()} [{msg.partition()}]')
-
+"""
 
 if __name__ == "__main__":
-    producer = SerializingProducer(kafka_producer_config)
     create_tables(connect, cursor)
     candidates = get_candidates()
-    print(len(candidates))
     if len(candidates) == 0:
-        trump_candidate = create_candidate_data(trump_data) # 3 is number of candidate
-        print(trump_data)
-        harris_candidate = create_candidate_data(harris_data) # 3 is number of candidate
-        print(trump_data)
-        insert_candidates(connect, cursor, trump_candidate)
-        insert_candidates(connect, cursor, harris_candidate)
+        candidate_list = add_candidates(trump_data, harris_data)
+        for candidate in candidate_list:
+            insert_candidates(connect, cursor, candidate)
 
+    for i in range(NUMBER_OF_VOTERS): 
 
-
-    for i in range(156000000): # 1000 is number of voters 
         voter_data = generate_voter_data()
-        print("voter_data", voter_data)
         insert_voters(connect, cursor, voter_data)
-        
-        producer.produce(
-            KAFKA_CONFIG["topic_voters"],# voters_topic
-            key=voter_data["voter_id"],
-            value=json.dumps(voter_data),   
-            on_delivery=delivery_report
-        )
-
+        kafka_producer(KAFKA_CONFIG['topic_voters'], voter_data)
         print('Produced voter {}, data: {}'.format(i, voter_data))
-        producer.flush()
+
 
 # kafka-topics --list --bootstrap-server broker:29092
 # kafka-console-consumer --topic topic_voters --bootstrap-server broker:9092 --from-beginning
-# kafka-topics --delete --topic vehicle_data --bootstrap-server broker:29092        
+# kafka-topics --delete --topic topic_votes --bootstrap-server broker:29092        
 # kafka-topics --delete --topic gps_data --bootstrap-server broker:29092        
 # kafka-topics --delete --topic traffic_data --bootstrap-server broker:29092        
 # kafka-topics --delete --topic weather_data --bootstrap-server broker:29092
