+++ kafka_service/docker-compose.kafka.yaml
@@ -1,0 +1,66 @@
+  version: '3.8'
+
+  services:
+    zookeeper:
+      image: confluentinc/cp-zookeeper:7.4.0
+      hostname: zookeeper
+      container_name: zookeeper
+      ports:
+        - "2181:2181"
+      environment:
+        ZOOKEEPER_CLIENT_PORT: 2181
+        ZOOKEEPER_TICK_TIME: 2000
+      healthcheck:
+        test: [ "CMD", "bash", "-c", "echo 'ruok' | nc localhost 2181" ]
+        interval: 10s
+        timeout: 5s
+        retries: 5
+      networks:
+        - voting-platform-nw
+        
+    broker:
+      image: confluentinc/cp-kafka:7.4.0
+      hostname: broker
+      container_name: broker
+      depends_on:
+        zookeeper:
+          condition: service_healthy
+      ports:
+        - "9092:9092"
+        - "9101:9101"
+      environment:
+        KAFKA_BROKER_ID: 1
+        KAFKA_ZOOKEEPER_CONNECT: 'zookeeper:2181'
+        KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: PLAINTEXT:PLAINTEXT,PLAINTEXT_HOST:PLAINTEXT
+        KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://broker:29092,PLAINTEXT_HOST://localhost:9092
+        KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
+        KAFKA_GROUP_INITIAL_REBALANCE_DELAY_MS: 0
+        KAFKA_CONFLUENT_LICENSE_TOPIC_REPLICATION_FACTOR: 1
+        KAFKA_CONFLUENT_BALANCER_TOPIC_REPLICATION_FACTOR: 1
+        KAFKA_TRANSACTION_STATE_LOG_MIN_ISR: 1
+        KAFKA_TRANSACTION_STATE_LOG_REPLICATION_FACTOR: 1
+        KAFKA_JMX_PORT: 9101
+        KAFKA_JMX_HOSTNAME: localhost
+      networks:
+        - voting-platform-nw
+      healthcheck:
+        test: [ "CMD", "bash", "-c", 'nc -z localhost 9092' ]
+        interval: 10s
+        timeout: 5s
+        retries: 5
+
+  networks:
+    voting-platform-nw:
+      external: true
+
+  # docker-compose -f docker-compose.kafka.yaml up --build -d 
+  # docker network inspect voting-platform-nw
+
+
+  # kafka-topics --list --bootstrap-server broker:29092
+  # kafka-console-consumer --topic topic_voters --bootstrap-server broker:9092 --from-beginning
+  # kafka-topics --delete --topic aggregated_turnout_by_location --bootstrap-server broker:29092        
+  # kafka-topics --delete --topic gps_data --bootstrap-server broker:29092        
+  # kafka-topics --delete --topic traffic_data --bootstrap-server broker:29092        
+  # kafka-topics --delete --topic weather_data --bootstrap-server broker:29092        
+  # kafka-topics --delete --topic emergency_data --bootstrap-server broker:29092
