+++ postgres_service/docker-compose.postgres.yaml
@@ -1,0 +1,43 @@
+version: '3.8'
+
+services:
+  postgres:
+    image: postgres:13
+    container_name: postgres
+    ports:
+      - "5432:5432"
+    environment:
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+      POSTGRES_DB: voting
+    healthcheck:
+      test: [ "CMD", "pg_isready", "-U", "postgres" ]
+      interval: 10s
+      timeout: 5s
+      retries: 5
+    volumes:
+      - "data-postgres:/var/lib/postgresql/data:rw"
+    networks:
+      - voting-platform-nw
+
+  pgadmin:
+    image: dpage/pgadmin4
+    container_name: pgadmin-container
+    environment:
+      - PGADMIN_DEFAULT_EMAIL=admin@admin.com
+      - PGADMIN_DEFAULT_PASSWORD=root
+    ports:
+      - "6868:80"
+    networks:
+      - voting-platform-nw
+
+volumes:
+  data-postgres:
+
+networks:
+  voting-platform-nw:
+    external: true  # Ensure this network is used as an external network
+
+# docker-compose -f docker-compose.postgres.yaml up --build
+# docker-compose -f docker-compose.postgres.yaml down  # Stop and remove containers, networks, etc.
+# docker-compose -f docker-compose.postgres.yaml up --build -d  # Rebuild and start containers
