--- requirements.txt
@@ -1,59 +1,0 @@
-altair==5.2.0
-attrs==23.1.0
-blinker==1.7.0
-cachetools==5.3.2
-certifi==2023.11.17
-charset-normalizer==3.3.2
-click==8.1.7
-confluent-kafka==2.3.0
-contourpy==1.2.0
-cycler==0.12.1
-Faker==20.1.0
-fonttools==4.46.0
-gitdb==4.0.11
-GitPython==3.1.40
-idna==3.6
-importlib-metadata==6.11.0
-importlib-resources==6.1.1
-Jinja2==3.1.2
-jsonschema==4.20.0
-jsonschema-specifications==2023.11.2
-kafka-python==2.0.2
-kiwisolver
-markdown-it-py==3.0.0
-MarkupSafe==2.1.3
-matplotlib==3.8.2
-mdurl==0.1.2
-numpy==1.26.2
-packaging==23.2
-pandas==2.1.3
-Pillow==10.1.0
-protobuf==3.20.1
-psycopg2-binary==2.9.9
-py4j==0.10.9.7
-pyarrow==14.0.1
-pydeck==0.8.1b0
-Pygments==2.17.2
-pyparsing==3.1.1
-pyspark==3.5.0
-python-dateutil==2.8.2
-pytz==2023.3.post1
-referencing==0.31.1
-requests==2.31.0
-rich==13.7.0
-rpds-py==0.13.2
-simplejson==3.19.2
-six==1.16.0
-smmap==5.0.1
-streamlit==1.29.0
-streamlit-autorefresh==1.0.1
-tenacity==8.2.3
-toml==0.10.2
-toolz==0.12.0
-tornado==6.4
-typing_extensions==4.8.0
-tzdata==2023.3
-tzlocal==5.2
-urllib3==2.1.0
-validators==0.22.0
-zipp==3.17.0
