+++ spark_service/docker-compose.spark.yaml
@@ -1,0 +1,41 @@
+version: '3.8'
+
+services:
+
+  spark-master:
+   image: bitnami/spark:latest
+   command: bin/spark-class org.apache.spark.deploy.master.Master
+   ports:
+     - "9090:8080"
+     - "7077:7077"
+   networks: 
+    - voting-platform-nw
+
+  spark-worker:
+   image: bitnami/spark:latest
+   command: bin/spark-class org.apache.spark.deploy.worker.Worker spark://spark-master:7077
+   depends_on:
+     - spark-master
+   environment:
+     SPARK_MODE: worker
+     SPARK_WORKER_CORES: 2
+     SPARK_WORKER_MEMORY: 1g
+     SPARK_MASTER_URL: spark://spark-master:7077
+   networks: 
+    - voting-platform-nw
+
+  spark-streaming:
+    build: .
+    container_name: spark-streaming
+    volumes:
+      - .:/app  
+    restart: unless-stopped  
+    networks:
+      - voting-platform-nw
+
+networks:
+  voting-platform-nw:
+    external: true
+
+# docker-compose -f docker-compose.spark.yaml up --build
+# docker-compose -f docker-compose.spark.yaml down
