+++ spark_service/requirements.txt
@@ -1,0 +1,1 @@
+pyspark==3.5.0
