+++ voting_serivce/Dockerfile
@@ -1,0 +1,7 @@
+
+FROM python:3.9-slim
+WORKDIR /app
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+COPY . .
+CMD ["python3", "main.py"]
